backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
PyPDF2==3.0.1
langchain==0.1.0
langchain-google-genai==0.0.6
langchain-community==0.0.10
google-generativeai==0.3.2
faiss-cpu==1.7.4
numpy==1.26.3
python-dotenv==1.0.0
pydantic==2.5.3
supabase==2.3.0
httpx>=0.24.0,<0.25.0
tiktoken==0.5.2
sentence-transformers==2.3.1
